+++ src/Utils/index.ts
@@ -118,15 +118,3 @@ class MyStorage {
 export const LS = new MyStorage(window.localStorage);
 export const SS = new MyStorage(window.sessionStorage);
 
-
-export const debounce = (fn: Function, delay: number = 800) => {
-  let timer: NodeJS.Timeout;
-  const debounceFn = function<T> (...args: T[]): any {
-    if (timer) window.clearTimeout(timer);
-    timer = setTimeout(() => {
-      fn.apply(debounceFn, args);
-      window.clearTimeout(timer);
-    }, delay);
-  }
-  return debounceFn;
-}
