+++ src/pages/user/index.tsx
@@ -1,37 +1,41 @@
-import React, { useReducer, createContext } from "react";
+import React, { useLayoutEffect } from "react";
 import SideBar from "@/common/sideBar/index";
 import { Layout, Card, Result, Button } from "@arco-design/web-react";
 import { Switch, Route } from "react-router-dom";
 import { routeMethod } from "@/route/getRoute";
 import { withRouter, RouteComponentProps } from "react-router";
 import { IconFaceSmileFill } from "@arco-design/web-react/icon";
-import { IUserAction, IUserState } from "./types";
-import { history } from "@/route";
 import TopHeader from "@/common/Header";
+import localStorageUtils from "@/Utils/localStorageUtils";
+import { useReduxDispatch } from '@/redux'
+import { useHistory, useLocation } from 'react-router-dom'
 import "./index.scss";
-import "./index.scss";
-export const UserContext = createContext((a: any) => a);
 
 const { Sider, Content, Header } = Layout;
 
-const initialState = {
-  flash: false,
-};
-
-const userReducer = (state: IUserState, action: IUserAction) => {
-  switch (action.type) {
-    case "flash":
-      return {
-        ...state,
-        flash: !state.flash,
-      };
-    default:
-      return state;
-  }
-};
-
 const User: React.FC<RouteComponentProps> = (props) => {
-  const [state, dispatch] = useReducer(userReducer, initialState);
+  const reduxDispatch = useReduxDispatch();
+  const location = useLocation();
+  const history = useHistory();
+
+  // 用户设置的路由拦截
+  useLayoutEffect(() => {
+    const { pathname } = location;
+    const data = localStorageUtils.get();
+    if (JSON.stringify(data) !== "{}") {
+      reduxDispatch({
+        type: 'userData/setData',
+        payload: {
+          login: true,
+          avatar: data.user.avatar,
+          name: data.user.name,
+        }
+      })
+    } else {
+      if (pathname.includes('/user')) history.push('/');
+    }
+  }, [])
+
 
   const renderContent = () => {
     return (
@@ -79,26 +83,24 @@ const User: React.FC<RouteComponentProps> = (props) => {
   };
 
   return (
-    <UserContext.Provider value={dispatch}>
-      <div className={`userPage`}>
-        <Layout className="user-box">
-          <Header>
-            <TopHeader />
-          </Header>
-          <Layout>
-            {/* 侧边栏 */}
-            <Sider className="home-page-sider-bar" collapsed={false}>
-              <SideBar />
-            </Sider>
-            <Content>
-              <Card className="card-box">
-                <Card className="content-box">{renderContent()}</Card>
-              </Card>
-            </Content>
-          </Layout>
+    <div className={`userPage`}>
+      <Layout className="user-box">
+        <Header>
+          <TopHeader />
+        </Header>
+        <Layout>
+          {/* 侧边栏 */}
+          <Sider className="home-page-sider-bar" collapsed={false}>
+            <SideBar />
+          </Sider>
+          <Content>
+            <Card className="card-box">
+              <Card className="content-box">{renderContent()}</Card>
+            </Card>
+          </Content>
         </Layout>
-      </div>
-    </UserContext.Provider>
+      </Layout>
+    </div>
   );
 };
